Changed parsers to use Array instead of InmutableArray. Removed unused parameters and methods on parsers. More work on tests. Ignored Test Projects on Code Coverage Analysis. Removed access modifiers on entities interfaces. Cleaned some code. Added Test.Shared project to centralize all common test objects.
