+++ src/Hyperar.HUM.Application.UnitTest/ChppFile/FileDownloadTaskFactoryTests.cs
@@ -4,7 +4,7 @@
     using System.Collections.Specialized;
     using Hyperar.HUM.Application.ChppFile.Download.Command.Interfaces;
     using Hyperar.HUM.Shared.Enums;
-    using Hyperar.HUM.TestShared;
+    using Hyperar.HUM.Test.Shared;
     using Microsoft.Extensions.DependencyInjection;
 
     public class FileDownloadTaskFactoryTests : IClassFixture<ServicesFixture>
