+++ src/Hyperar.HUM.Application.UnitTest/ChppFile/FileExtractStrategyFactoryTests.cs
@@ -4,7 +4,7 @@
     using Hyperar.HUM.Application.ChppFile.Download.Command.Interfaces;
     using Hyperar.HUM.Application.ChppFile.Download.Command.Strategies.Extract;
     using Hyperar.HUM.Shared.Enums;
-    using Hyperar.HUM.TestShared;
+    using Hyperar.HUM.Test.Shared;
     using Microsoft.Extensions.DependencyInjection;
 
     public class FileExtractStrategyFactoryTests : IClassFixture<ServicesFixture>
