+++ src/Hyperar.HUM.Application/ChppFile/Download/Command/Strategies/Parse/ManagerCompendiumParser.cs
@@ -28,8 +28,7 @@
                 userId,
                 fetchetDate,
                 await ReadManagerNodeAsync(
-                    xmlReader,
-                    cancellationToken));
+                    xmlReader));
         }
 
         private static async Task<Currency> ReadCurrencyNodeAsync(XmlReader xmlReader)
@@ -45,7 +44,7 @@
             return currency;
         }
 
-        private static async Task<IEnumerable<string>> ReadLastLoginsNodeAsync(XmlReader xmlReader)
+        private static async Task<string[]> ReadLastLoginsNodeAsync(XmlReader xmlReader)
         {
             await xmlReader.ReadAsync();
 
@@ -59,7 +58,7 @@
 
             await xmlReader.ReadAsync();
 
-            return lastLogins;
+            return lastLogins.ToArray();
         }
 
         private static async Task<League> ReadLeagueNodeAsync(XmlReader xmlReader)
@@ -76,7 +75,7 @@
             return league;
         }
 
-        private static async Task<Manager> ReadManagerNodeAsync(XmlReader xmlReader, CancellationToken cancellationToken)
+        private static async Task<Manager> ReadManagerNodeAsync(XmlReader xmlReader)
         {
             await xmlReader.ReadAsync();
 
@@ -88,37 +87,31 @@
                     xmlReader),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.Language,
-                    cancellationToken),
+                    NodeName.Language),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.Country,
-                    cancellationToken),
+                    NodeName.Country),
                 await ReadCurrencyNodeAsync(
                     xmlReader),
                 await ReadTeamsNodeAsync(
-                    xmlReader,
-                    cancellationToken),
+                    xmlReader),
                 await ReadNullableIdNameListNodeAsync(
                     xmlReader,
                     NodeName.NationalTeamCoach,
-                    NodeName.NationalTeam,
-                    cancellationToken),
+                    NodeName.NationalTeam),
                 await ReadNullableIdNameListNodeAsync(
                     xmlReader,
                     NodeName.NationalTeamAssistant,
-                    NodeName.NationalTeam,
-                    cancellationToken),
+                    NodeName.NationalTeam),
                 await ReadAvatarNodeAsync(
-                    xmlReader,
-                    cancellationToken));
+                    xmlReader));
 
             await xmlReader.ReadAsync();
 
             return manager;
         }
 
-        private static async Task<Team> ReadTeamNodeAsync(XmlReader xmlReader, CancellationToken cancellationToken)
+        private static async Task<Team> ReadTeamNodeAsync(XmlReader xmlReader)
         {
             await xmlReader.ReadAsync();
 
@@ -127,34 +120,28 @@
                 (await xmlReader.ReadValueAsync()).AsString(),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.Arena,
-                    cancellationToken),
+                    NodeName.Arena),
                 await ReadLeagueNodeAsync(
                     xmlReader),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.Country,
-                    cancellationToken),
+                    NodeName.Country),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.LeagueLevelUnit,
-                    cancellationToken),
+                    NodeName.LeagueLevelUnit),
                 await ReadIdNameNodeAsync(
                     xmlReader,
-                    NodeName.Region,
-                    cancellationToken),
+                    NodeName.Region),
                 await ReadYouthTeamNodeAsync(
-                    xmlReader,
-                    cancellationToken));
+                    xmlReader));
 
             await xmlReader.ReadAsync();
 
             return team;
         }
 
-        private static async Task<IEnumerable<Team>> ReadTeamsNodeAsync(
-            XmlReader xmlReader,
-            CancellationToken cancellationToken)
+        private static async Task<Team[]> ReadTeamsNodeAsync(
+            XmlReader xmlReader)
         {
             if (!xmlReader.CheckNode(NodeName.Teams))
             {
@@ -173,16 +160,15 @@
             {
                 teamList.Add(
                     await ReadTeamNodeAsync(
-                        xmlReader,
-                        cancellationToken));
+                        xmlReader));
             }
 
             await xmlReader.ReadAsync();
 
-            return teamList;
+            return teamList.ToArray();
         }
 
-        private static async Task<YouthTeam?> ReadYouthTeamNodeAsync(XmlReader xmlReader, CancellationToken cancellationToken)
+        private static async Task<YouthTeam?> ReadYouthTeamNodeAsync(XmlReader xmlReader)
         {
             YouthTeam? youthTeam = null;
 
@@ -195,8 +181,7 @@
                    (await xmlReader.ReadValueAsync()).AsString(),
                    await ReadNullableIdNameNodeAsync(
                        xmlReader,
-                       NodeName.YouthLeague,
-                       cancellationToken));
+                       NodeName.YouthLeague));
             }
 
             await xmlReader.ReadAsync();
