+++ src/Hyperar.HUM.Application/ChppFile/Download/Command/Strategies/Parse/XmlParserBase.cs
@@ -9,7 +9,7 @@
 
     public abstract class XmlParserBase
     {
-        protected static async Task<Avatar?> ReadAvatarNodeAsync(XmlReader xmlReader, CancellationToken cancellationToken)
+        protected static async Task<Avatar?> ReadAvatarNodeAsync(XmlReader xmlReader)
         {
             // Reads opening node.
             await xmlReader.ReadAsync();
@@ -17,8 +17,7 @@
             var result = new Avatar(
                 (await xmlReader.ReadValueAsync()).AsString(),
                 await ReadLayerNodes(
-                    xmlReader,
-                    cancellationToken));
+                    xmlReader));
 
             // Reads closing node.
             await xmlReader.ReadAsync();
@@ -26,32 +25,9 @@
             return result;
         }
 
-        protected static async Task<IEnumerable<IdName>> ReadIdNameListNodeAsync(
-            XmlReader xmlReader,
-            string expectedName,
-            string expectedChildName,
-            CancellationToken cancellationToken)
-        {
-            if (!xmlReader.CheckNode(expectedName))
-            {
-                throw new BusinessException(
-                    string.Format(
-                        Globalization.ErrorMessages.InvalidXmlElement,
-                        expectedName,
-                        xmlReader.Name));
-            }
-
-            return await ReadIdNameListNodeInternalAsync(
-                xmlReader,
-                expectedName,
-                expectedChildName,
-                cancellationToken);
-        }
-
         protected static async Task<IdName> ReadIdNameNodeAsync(
             XmlReader xmlReader,
-            string expectedName,
-            CancellationToken cancellationToken)
+            string expectedName)
         {
             if (!xmlReader.CheckNode(expectedName))
             {
@@ -73,11 +49,10 @@
             return node;
         }
 
-        protected static async Task<IEnumerable<IdName>?> ReadNullableIdNameListNodeAsync(
+        protected static async Task<IdName[]?> ReadNullableIdNameListNodeAsync(
             XmlReader xmlReader,
             string expectedName,
-            string expectedChildName,
-            CancellationToken cancellationToken)
+            string expectedChildName)
         {
             if (!xmlReader.CheckNode(expectedName))
             {
@@ -86,15 +61,12 @@
 
             return await ReadIdNameListNodeInternalAsync(
                 xmlReader,
-                expectedName,
-                expectedChildName,
-                cancellationToken);
+                expectedChildName);
         }
 
         protected static async Task<IdName?> ReadNullableIdNameNodeAsync(
             XmlReader xmlReader,
-            string expectedName,
-            CancellationToken cancellationToken)
+            string expectedName)
 
         {
             if (xmlReader.IsEmptyElement)
@@ -121,11 +93,9 @@
             return node;
         }
 
-        private static async Task<IEnumerable<IdName>> ReadIdNameListNodeInternalAsync(
+        private static async Task<IdName[]> ReadIdNameListNodeInternalAsync(
             XmlReader xmlReader,
-            string expectedName,
-            string expectedChildName,
-            CancellationToken cancellationToken)
+            string expectedChildName)
         {
             var nodes = new List<IdName>();
 
@@ -136,16 +106,15 @@
                 nodes.Add(
                     await ReadIdNameNodeAsync(
                         xmlReader,
-                        expectedChildName,
-                        cancellationToken));
+                        expectedChildName));
             }
 
             await xmlReader.ReadAsync();
 
-            return nodes;
+            return nodes.ToArray();
         }
 
-        private static async Task<Layer> ReadLayerNodeAsync(XmlReader xmlReader, CancellationToken cancellationToken)
+        private static async Task<Layer> ReadLayerNodeAsync(XmlReader xmlReader)
         {
             var x = xmlReader.GetAttribute(NodeName.X).AsInt();
             var y = xmlReader.GetAttribute(NodeName.Y).AsInt();
@@ -162,7 +131,7 @@
             return result;
         }
 
-        private static async Task<IEnumerable<Layer>> ReadLayerNodes(XmlReader xmlReader, CancellationToken cancellationToken)
+        private static async Task<Layer[]> ReadLayerNodes(XmlReader xmlReader)
         {
             var nodes = new List<Layer>();
 
@@ -170,11 +139,10 @@
             {
                 nodes.Add(
                     await ReadLayerNodeAsync(
-                        xmlReader,
-                        cancellationToken));
+                        xmlReader));
             }
 
-            return nodes;
+            return nodes.ToArray();
         }
     }
 }
