+++ src/Hyperar.HUM.ChppApiClient.IntegrationTest/Hyperar.HUM.ChppApiClient.IntegrationTest.csproj
@@ -18,6 +18,7 @@
         <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.1" />
         <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.1" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
+        <PackageReference Include="Moq" Version="4.20.72" />
         <PackageReference Include="xunit" Version="2.9.3" />
         <PackageReference Include="xunit.runner.visualstudio" Version="3.0.1">
             <PrivateAssets>all</PrivateAssets>
@@ -27,7 +28,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Hyperar.HUM.ChppApiClient\Hyperar.HUM.ChppApiClient.csproj" />
-        <ProjectReference Include="..\Hyperar.HUM.TestShared\Hyperar.HUM.TestShared.csproj" />
+        <ProjectReference Include="..\Hyperar.HUM.Test.Shared\Hyperar.HUM.Test.Shared.csproj" />
     </ItemGroup>
 
     <ItemGroup>
@@ -38,6 +39,10 @@
         <None Update="appSettings.json">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
+    </ItemGroup>
+
+    <ItemGroup>
+        <AssemblyAttribute Include="System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverageAttribute" />
     </ItemGroup>
 
 </Project>
