+++ src/Hyperar.HUM.ChppApiClient/Hyperar.HUM.ChppApiClient.csproj
@@ -11,7 +11,7 @@
         <PackageReference Include="Hyperar.OAuthCore" Version="2.0.1" />
         <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.1" />
         <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.1" />
-        <PackageReference Include="WireMock.Net" Version="1.7.0" />
+        <PackageReference Include="WireMock.Net" Version="1.7.1" />
     </ItemGroup>
 
     <ItemGroup>
