+++ src/Hyperar.HUM.Shared/Models/Chpp/ManagerCompendium/Manager.cs
@@ -1,17 +1,15 @@
 ﻿namespace Hyperar.HUM.Shared.Models.Chpp.ManagerCompendium
 {
-    using System.Collections.Generic;
-
     public record Manager(
         long UserId,
         string LoginName,
         string SupporterTier,
-        IEnumerable<string> LastLogins,
+        string[] LastLogins,
         IdName Language,
         IdName Country,
         Currency Currency,
-        IEnumerable<Team> Teams,
-        IEnumerable<IdName>? NationalTeamCoach,
-        IEnumerable<IdName>? NationalTeamAssistant,
+        Team[] Teams,
+        IdName[]? NationalTeamCoach,
+        IdName[]? NationalTeamAssistant,
         Avatar? Avatar);
 }
