+++ src/Hyperar.HUM.Shared/Models/Chpp/WorldDetails/Country.cs
@@ -1,7 +1,5 @@
 ﻿namespace Hyperar.HUM.Shared.Models.Chpp.WorldDetails
 {
-    using System.Collections.Generic;
-
     public record Country(
         bool Available,
         long? CountryId,
@@ -11,5 +9,5 @@
         string? CountryCode,
         string? DateFormat,
         string? TimeFormat,
-        IEnumerable<IdName>? RegionList);
+        IdName[]? RegionList);
 }
