+++ src/Hyperar.HUM.Shared/Models/Chpp/WorldDetails/HattrickData.cs
@@ -1,7 +1,6 @@
 ﻿namespace Hyperar.HUM.Shared.Models.Chpp.WorldDetails
 {
     using System;
-    using System.Collections.Generic;
     using Hyperar.HUM.Shared.Models.Chpp.Interfaces;
 
     public record HattrickData(
@@ -9,5 +8,5 @@
         decimal Version,
         long UserId,
         DateTime FetchedDate,
-        IEnumerable<League> LeagueList) : XmlFileBase(FileName, Version, UserId, FetchedDate), IXmlFileBase;
+        League[] LeagueList) : XmlFileBase(FileName, Version, UserId, FetchedDate), IXmlFileBase;
 }
