+++ src/Hyperar.HUM.Test.Shared/Hyperar.HUM.Test.Shared.csproj
@@ -8,11 +8,16 @@
 
     <ItemGroup>
       <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="9.0.1" />
+      <PackageReference Include="Moq" Version="4.20.72" />
     </ItemGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Hyperar.HUM.Application\Hyperar.HUM.Application.csproj" />
       <ProjectReference Include="..\Hyperar.HUM.Infrastructure\Hyperar.HUM.Infrastructure.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <AssemblyAttribute Include="System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverageAttribute" />
     </ItemGroup>
 
 </Project>
